Refactor scroll logic to own helper & fix N+1 false positives
- Scroll logic is also used by the search spec here: spec/system/learnings_search_spec.rb, hence extracting it makes the code DRY & more easily reusable
