+++ spec/system/learnings_infinite_scroll_spec.rb
@@ -12,6 +12,7 @@ RSpec.describe 'Learnings Infinite Scroll', type: :system do
 
   context 'with multiple pages of learnings' do
     before do
+      Prosopite.pause
       50.times do |i|
         create(:learning,
                lesson: "Learning #{i + 1}",
@@ -20,6 +21,7 @@ RSpec.describe 'Learnings Infinite Scroll', type: :system do
                created_at: Time.zone.now - (49 - i).minutes,
                updated_at: Time.zone.now - (49 - i).minutes)
       end
+      Prosopite.resume
     end
 
     it 'loads learnings progressively as user scrolls' do
@@ -36,22 +38,6 @@ RSpec.describe 'Learnings Infinite Scroll', type: :system do
 
       # Scroll until we see the oldest item (all pages loaded)
       scroll_until_content_appears('Learning 1')
-    end
-
-    # Turbo lazy loading relies on IntersectionObserver which is unreliable in headless Chrome.
-    # This helper scrolls repeatedly until the expected content appears.
-    def scroll_until_content_appears(content, max_attempts: 2)
-      max_attempts.times do
-        return if page.has_content?(content)
-
-        page.scroll_to(:bottom)
-        page.execute_script(<<~JS)
-          const frame = document.querySelector('turbo-frame[loading=lazy]');
-          if (frame) { frame.loading = 'eager'; frame.reload(); }
-        JS
-      end
-
-      expect(page).to have_content(content)
     end
   end
 end
